improtinv randmt32 and dice modules

+++ usr/lib/zero/randmt32.h
@@ -5,7 +5,7 @@
 
 #define RANDMT32_MAX 0xffffffff
 
-void          srandmt32(int32_t seed);
+void          srandmt32(unsigned long seed);
 unsigned long randmt32(void);
 
 #endif /* __ZERO_RANDMT32_H__ */
